more tweet support

+++ master/tweetMode.py
@@ -29,7 +29,7 @@ brewlog=open("ipc/brewlog-id").read().rstrip()
 ## GRAIN
 ##
 
-if os.path.exists("ipc/activityDough") and not os.path.exists("ipc/tweet-grainbill"):
+if not os.path.exists("ipc/swFerm") and os.path.exists("ipc/activityDough") and not os.path.exists("ipc/tweet-grainbill"):
 	flag=open("ipc/tweet-grainbill","w")
 	flag.close()	
 	con=mysql.connector.connect(host="192.168.1.13",user='brewerslab',password='beer',database="brewerslab")
@@ -44,7 +44,7 @@ if os.path.exists("ipc/activityDough") and not os.path.exists("ipc/tweet-grainbi
 ## HOPS
 ##
 ## FWH
-if os.path.exists("ipc/activityHltSparge") and not os.path.exists("ipc/swSparge") and not os.path.exists("ipc/tweet-fwh"):
+if not os.path.exists("ipc/swFerm") and os.path.exists("ipc/activityHltSparge") and not os.path.exists("ipc/swSparge") and not os.path.exists("ipc/tweet-fwh"):
 	flag=open("ipc/tweet-fwh","w")
 	flag.close()	
 	con=mysql.connector.connect(host="192.168.1.13",user='brewerslab',password='beer',database="brewerslab")
@@ -58,7 +58,7 @@ if os.path.exists("ipc/activityHltSparge") and not os.path.exists("ipc/swSparge"
 
 
 ## COPPER 
-if os.path.exists("activityReachedBoil") and not os.path.exists("ipc/tweet-copperhop"):
+if not os.path.exists("ipc/swFerm") and os.path.exists("activityReachedBoil") and not os.path.exists("ipc/tweet-copperhop"):
 	flag=open("ipc/tweet-copperhop","w")
 	flag.close()	
 	con=mysql.connector.connect(host="192.168.1.13",user='brewerslab',password='beer',database="brewerslab")
@@ -71,9 +71,39 @@ if os.path.exists("activityReachedBoil") and not os.path.exists("ipc/tweet-coppe
 
 
 #
+# Aroma
+#
+if not os.path.exists("ipc/swFerm") and os.path.exists("ipc/activityAromaHops") and os.path.exists("ipc/swBoil") and not os.path.exists("ipc/tweet-aroma"):
+	flag=open("ipc/tweet-aroma","w")
+	flag.close()	
+	con=mysql.connector.connect(host="192.168.1.13",user='brewerslab',password='beer',database="brewerslab")
+	cursor=con.cursor()
+	cursor.execute("select recipe,ingredient,qty,unit,hopAddAt from gIngredients,gBrewlogs where gBrewlogs.recipe=gIngredients.recipename and brewlog='%s' and ingredientType='hops' AND hopAddAt >5 AND hopAddAt < 20" %(brewlog))
+	for row in cursor:
+		(recipename,ingredient,qty,unit,hopAddAt)=row
+		print row
+		doTweet("added %s %s of %s Aroma Hops #hopbill #brewerslab #%s" %(qty,unit,ingredient,re.compile("[^A-Za-z0-9]").sub('',recipename)))
+
+
+#
+# Aroma
+#
+if not os.path.exists("ipc/swFerm") and os.path.exists("ipc/activityFinishingHops") and os.path.exists("ipc/swBoil") and not os.path.exists("ipc/tweet-finishhop"):
+	flag=open("ipc/tweet-finishhop","w")
+	flag.close()	
+	con=mysql.connector.connect(host="192.168.1.13",user='brewerslab',password='beer',database="brewerslab")
+	cursor=con.cursor()
+	cursor.execute("select recipe,ingredient,qty,unit,hopAddAt from gIngredients,gBrewlogs where gBrewlogs.recipe=gIngredients.recipename and brewlog='%s' and ingredientType='hops' AND hopAddAt >3 AND hopAddAt < 10" %(brewlog))
+	for row in cursor:
+		(recipename,ingredient,qty,unit,hopAddAt)=row
+		print row
+		doTweet("added %s %s of %s Finishing Hops #hopbill #brewerslab #%s" %(qty,unit,ingredient,re.compile("[^A-Za-z0-9]").sub('',recipename)))
+
+
+#
 # Flameout
 #
-if os.path.exists("ipc/activityFlameoutHops") and not os.path.exists("ipc/swBoil") and not os.path.exists("ipc/tweet-flameout"):
+if not os.path.exists("ipc/swFerm") and os.path.exists("ipc/activityFlameoutHops") and not os.path.exists("ipc/swBoil") and not os.path.exists("ipc/tweet-flameout"):
 	flag=open("ipc/tweet-flameout","w")
 	flag.close()	
 	con=mysql.connector.connect(host="192.168.1.13",user='brewerslab',password='beer',database="brewerslab")
@@ -89,11 +119,12 @@ if os.path.exists("ipc/activityFlameoutHops") and not os.path.exists("ipc/swBoil
 #
 # Whirlpool
 #
-if os.path.exists("ipc/activityWirlpool") and not os.path.exists("ipc/tweet-whirlpoolhop"):
+if not os.path.exists("ipc/swFerm") and os.path.exists("ipc/activityWhirlpool") and not os.path.exists("ipc/tweet-whirlpoolhop"):
 	flag=open("ipc/tweet-whirlpoolhop","w")
 	flag.close()	
 	con=mysql.connector.connect(host="192.168.1.13",user='brewerslab',password='beer',database="brewerslab")
 	cursor=con.cursor()
+	print "select recipe,ingredient,qty,unit,hopAddAt from gIngredients,gBrewlogs where gBrewlogs.recipe=gIngredients.recipename and brewlog='%s' and ingredientType='hops' AND hopAddAt >0.001 AND hopAddAt < 0.003" %(brewlog)
 	cursor.execute("select recipe,ingredient,qty,unit,hopAddAt from gIngredients,gBrewlogs where gBrewlogs.recipe=gIngredients.recipename and brewlog='%s' and ingredientType='hops' AND hopAddAt >0.001 AND hopAddAt < 0.003" %(brewlog))
 	for row in cursor:
 		(recipename,ingredient,qty,unit,hopAddAt)=row
@@ -102,4 +133,35 @@ if os.path.exists("ipc/activityWirlpool") and not os.path.exists("ipc/tweet-whir
 
 
 
+
+
+# Boil Volume
+if os.path.exists("ipc/swBoil") and not os.path.exists("ipc/tweet-boilvol"):
+	con=mysql.connector.connect(host="192.168.1.13",user='brewerslab',password='beer',database="brewerslab")
+	cursor=con.cursor()
+	cursor.execute("select gBrewlogs.recipe,fieldVal from gField,gBrewlogs WHERE gField.brewlog='%s' AND gBrewlogs.brewlog = gField.brewlog AND fieldKey='preBoilVolume' and length(fieldVal)>1" %(brewlog))
+	for row in cursor:
+		flag=open("ipc/tweet-boilvol","w")
+		flag.close()	
+		(recipename,fieldVal)=row
+		doTweet("currently boiling %sL #brewerslab #%s" %(fieldVal,re.compile("[^A-Za-z0-9]").sub('',recipename)))
+
+# Original Gravity
+if os.path.exists("ipc/swFerm") and not os.path.exists("ipc/tweet-og"):
+	og=None
+	con=mysql.connector.connect(host="192.168.1.13",user='brewerslab',password='beer',database="brewerslab")
+	cursor=con.cursor()
+	cursor.execute("select gBrewlogs.recipe,fieldVal from gField,gBrewlogs WHERE gField.brewlog='%s' AND gBrewlogs.brewlog = gField.brewlog AND fieldKey='og' and length(fieldVal)>1" %(brewlog))
+	for row in cursor:
+		og=fieldVal
+
+	if og:
+		con=mysql.connector.connect(host="192.168.1.13",user='brewerslab',password='beer',database="brewerslab")
+		cursor=con.cursor()
+		cursor.execute("select gBrewlogs.recipe,fieldVal from gField,gBrewlogs WHERE gField.brewlog='%s' AND gBrewlogs.brewlog = gField.brewlog AND fieldKey='postboilvol' and length(fieldVal)>1" %(brewlog))
+		for row in cursor:
+			flag=open("ipc/tweet-og","w")
+			flag.close()	
+			(recipename,fieldVal)=row
+			doTweet("%sL of wort in the fermenter - %.4f OG #brewerslab #%s" %(fieldVal,float(og),re.compile("[^A-Za-z0-9]").sub('',recipename)))
